+++ Rakefile
@@ -9,4 +9,4 @@ require 'metasploit/framework/require'
 Metasploit::Framework::Require.optionally_active_record_railtie
 
 Metasploit::Framework::Application.load_tasks
-Metasploit::Framework::Spec::Constants.define_task
+Metasploit::Framework::Spec::Constants::Suite.define_task
