+++ lib/metasploit/framework/spec/constants.rb
@@ -3,10 +3,16 @@ require 'msf/core/modules'
 # Monitor constants created by module loading to ensure that the loads in one example don't interfere with the
 # assertions in another example.
 module Metasploit::Framework::Spec::Constants
+  extend ActiveSupport::Autoload
+
+  autoload :Suite
+
+  #
+  # CONSTANTS
+  #
+
   # Regex parsing loaded module constants
   LOADED_MODULE_CHILD_CONSTANT_REGEXP = /^Mod(?<unpacked_full_name>[0-9a-f]+)$/
-  # Path to log holding leaked constants from last spec run.
-  LOG_PATHNAME = Pathname.new('log/leaked-constants.log')
   # The parent namespace child_constant_name that can have children added when loading modules.
   PARENT_CONSTANT = Msf::Modules
   # Constant names under {PARENT_CONSTANT} that can persist between specs because they are part of the loader library
@@ -18,59 +24,6 @@ module Metasploit::Framework::Spec::Constants
     Namespace
     VersionCompatibilityError
   }.map(&:to_sym)
-
-  # Configures after(:suite) callback for RSpec to check for leaked constants.
-  def self.configure!
-    unless @configured
-      RSpec.configure do |config|
-        config.after(:suite) do
-          count = 0
-
-          LOG_PATHNAME.open('w') do |f|
-            count = ::Metasploit::Framework::Spec::Constants.each { |child_name|
-              f.puts child_name
-            }
-          end
-
-          if count > 0
-            $stderr.puts "#{count} #{'constant'.pluralize(count)} leaked under #{PARENT_CONSTANT}.  " \
-                         "See #{LOG_PATHNAME} for details."
-          else
-            LOG_PATHNAME.delete
-          end
-        end
-      end
-
-      @configured = true
-    end
-  end
-
-  # Adds action to `spec` task so that `rake spec` fails if `log/leaked-constants.log` exists after printing out the
-  # leaked constants.
-  #
-  # @return [void]
-  def self.define_task
-    Rake::Task.define_task(:spec) do
-      if LOG_PATHNAME.exist?
-        $stderr.puts "Leaked constants detected under #{PARENT_CONSTANT}:"
-
-        LOG_PATHNAME.open do |f|
-          f.each_line do |line|
-            constant_name = line.strip
-            full_name = self.full_name(constant_name)
-
-            if full_name
-              formatted_full_name = " # #{full_name}"
-            end
-
-            $stderr.puts "  #{constant_name}#{formatted_full_name}"
-          end
-        end
-
-        exit 1
-      end
-    end
-  end
 
   # Yields each child_constant_name under {PARENT_CONSTANT}.
   #
