+++ spec/spec_helper.rb
@@ -54,4 +54,4 @@ RSpec.configure do |config|
   config.use_transactional_fixtures = true
 end
 
-Metasploit::Framework::Spec::Constants.configure!
+Metasploit::Framework::Spec::Constants::Suite.configure!
